# Single allocation `flattenChunks`