--- a/utils.go
+++ b/utils.go
@@ -1,8 +1,13 @@
 package rsmt2d
 
 func flattenChunks(chunks [][]byte) []byte {
-	flattened := chunks[0]
-	for _, chunk := range chunks[1:] {
+	length := 0
+	for _, chunk := range chunks {
+		length += len(chunk)
+	}
+
+	flattened := make([]byte, 0, length)
+	for _, chunk := range chunks {
 		flattened = append(flattened, chunk...)
 	}